# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "区块链房地产代币化系统 API",
    "version": "1.0.0",
    "description": "测试配置"
  },
  "paths": {
...
    "/api/v1/core/real-estate-facade/register-property": {
      "post": {
        "summary": "注册新房产",
        "description": "注册新房产并发行代币",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "country",
                  "metadataURI",
                  "initialSupply",
                  "tokenName",
                  "tokenSymbol"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "description": "房产ID"
                  },
                  "country": {
                    "type": "string",
                    "description": "国家"
                  },
                  "metadataURI": {
                    "type": "string",
                    "description": "元数据URI"
                  },
                  "initialSupply": {
                    "type": "string",
                    "description": "初始代币供应量"
                  },
                  "tokenName": {
                    "type": "string",
                    "description": "代币名称"
                  },
                  "tokenSymbol": {
                    "type": "string",
                    "description": "代币符号"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "房产注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/real-estate-facade/property/{propertyId}": {
      "get": {
        "summary": "获取房产信息",
        "description": "获取指定房产的详细信息",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "房产ID"
          }
        ],
        "responses": {
          "200": {
            "description": "房产信息获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "房产不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/real-estate-facade/update-property-status": {
      "put": {
        "summary": "更新房产状态",
        "description": "更新指定房产的状态",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "status"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "description": "房产ID"
                  },
                  "status": {
                    "type": "integer",
                    "description": "房产状态",
                    "enum": [
                      0,
                      1,
                      2,
                      3,
                      4,
                      5
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "房产状态更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/real-estate-facade/list-properties": {
      "get": {
        "summary": "获取房产列表",
        "description": "获取系统中的所有房产列表",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1,
                2,
                3,
                4,
                5
              ]
            },
            "description": "可选的状态过滤"
          }
        ],
        "responses": {
          "200": {
            "description": "房产列表获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/real-estate-facade/create-sell-order": {
      "post": {
        "summary": "创建卖单",
        "description": "创建房产代币卖单",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tokenAddress",
                  "amount",
                  "price"
                ],
                "properties": {
                  "tokenAddress": {
                    "type": "string",
                    "description": "代币合约地址"
                  },
                  "amount": {
                    "type": "string",
                    "description": "卖出金额"
                  },
                  "price": {
                    "type": "string",
                    "description": "单价"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "卖单创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/real-estate-facade/buy-tokens": {
      "post": {
        "summary": "购买代币",
        "description": "购买指定卖单的代币",
        "tags": [
          "RealEstateFacade"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderId"
                ],
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "卖单ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "代币购买成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}